Add visual feedback when resizing
